+++ treepoem/__init__.py
@@ -112,7 +112,9 @@ def _get_ghostscript_binary():
 
 
 def _encode(data):
-    return codecs.encode(data.encode('utf8'), 'hex_codec').decode('ascii')
+    if isinstance(data, str):
+        data = data.encode('utf8')
+    return codecs.encode(data, 'hex_codec').decode('ascii')
 
 
 def _format_options(options):